requirements.txt
asgiref==3.6.0
Django==4.1.4
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
social-auth-app-django==5.0.0
django-ranged-response-0.2.0 
django-simple-captcha-0.5.17